+++ CHANGELOG.md
@@ -10,11 +10,14 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Added
 
 - **doc**: Add select stories
+- **formly-field-tooltip**: Add possibility to customize the tooltip : content, icon, color, size.
+
 
 ### Changed
 
 - **grid**: add refresh sort when data changes
 - **doc**: migrate angular integration tutorial to storybook
+- **formly-field-tooltip**: please use the new formly templateOptions for tooltip :  `tooltip: { content: 'string' }` instead of `tooltip: 'string'`
 
 ### Deprecated